Java-DSA-Project 1.1
 Deadlock Detection in Scheduling
You are tasked with designing and implementing a code that detects deadlock scenarios 
in a WAIT-FOR graph used in process scheduling. Deadlocks occur when a circular 
dependency forms among multiple nodes in the graph, leading to a situation where none 
of the nodes can proceed.
Conditions
a. You will create a class to represent a Process and implement a graph of Processes 
(usually such a graph is directed and is called a Wait-For graph)
i. Process will have a name, process ID as fields (among other fields that
you wish to use in your code)
b. You will then detect cycle in the Graph on projects